Add move ordering and pruning based on distance to black win
For each possible way to win, including rotations, we compute the
number of black moves required for a black victory.  This helps
in two ways: (1) it provides a lower bound on the depth needed to
see a win for pruning purposes and (2) we can reorder moves to
try to maximize (black) or minimize (white) closeness to a black win.