Add distance-based move ordering and pruning to the rotation abstracted engine
This information is somewhat crude, since it measures distance to potential
wins allowing black to make any arbitrary rotations.  For move ordering this
makes sense, as the ordering isn't allowed to depend on rotation, but for
pruning it'd be nice if there was a fast rotated-aware method.  Unfortunately,
I think too bits are involved to make such a thing practical.