Add expensive diagnostic tracing code
If an inconsistency is detected during the search (currently only possible
during transposition table store), the tracing code will automatically
reduce the error to a minimal counterexample (a invalid position which
depends only on valid results).  Tracing is disabled by default; uncomment
the nonnull definition of TRACE in trace.h to enable it.