all_boards: Add a stable method of enumerating nearly standardized boards
The ApproximateBoards class provides a slightly redundant enumeration of
symmetry reduced boards, using an alternate standardization that is stable
under stone placement.  Specifically, we divide the space of boards into
sections based on the number of stones of each color in each quadrant,
and standardize the space of sections.  The set of boards in one section
forms a simple four dimensional grid.  The resulting enumeration adds only
10% redundancy over full superstandardization.

Critically, a single section of n stone positions can be computed using
data from only 4 n+1 stone sections (that given by adding a stone to
each of the four quadrants).