Adjust endgame computation to support reflection standardization
Instead of storing rotation minimal quadrants in sorted order per bucket,
we now put reflection pairs next to each other, and move all such pairs
to the front of the bucket.  As long as the bucket size is even, this means
that reflections can be performed bucket-wise, which recovers a lost
factor of (almost) 2 in performance.