Ponder strategy for strongly solving pentago
It looks like forward search can handle 18 stones through to the end
of the game, so we can compute perfect play from any position as long
as we store endgame database from 0 to 17 stones.  This will consume
around 2 TB, which is not bad.  I can afford to buy a 2 TB disk,
presumably.