mpi/partition: Even out distribution of blocks and lines
It's probably bad if a process gets a huge number of tiny blocks
and lines, since it will create a focal point for a large number
of messages.  Therefore, artificially inflate the cost of small
blocks and lines when dividing work between processes.  These
penalties can be adjusted in the line_penalty function.