mid: Implement downstream retrograde analysis
Start at a root board with 18 or more stones, the mid solver performs
an exhaustive backwards traversal of all downstream positions, keeping
track of rotations only via super_t.  No standardization is performed,
since 18 stone positions are likely to be quite rigid.

Currently this takes between 15.4 and 20 s for an 18 stone boards.
Wonderfully, this time should be completely invariant to the board.