Remove pkgcloud dependency
We only need a tiny portion of the interface to Rackspace: read-only
access to cloud files.  It turns out we can compress the relevant code
into less than 100 lines of clean Javascript.  This avoids pkgcloud as a
dependency, which is nice since pkgcloud is (1) huge, (2) messily
object-oriented, and (3) depends on a bunch of old versions of npm
packages since (4) it's unmaintained.