Switch midengine to side-based representation to reduce code size
We're blowing a bunch of code size on converting between packed ternary
and unpacked binary representations.  This eliminates that overhead by
representing high_board_t in terms of side_t, not board_t.

Other changes: high_board_t now stores the full ply count rather than
just middle, so that turn and such is ~free, and midsolve was inlined
into the wasm codepath to avoid extra code size due to copying piles
(I'll remove more of that in a future pass).

This shrinks mid.wasm from 16,931 to 13,975 bytes.  mid_worker.js grows
a bit due to the pack/unpack logic (from 3341 to 3738 bytes), but that
can be streamlined away by giving the worker a better interface.