mid: Use aggressive/passive to cut memory by half
I'm seeing malloc failures on my iPhone, so it's time to break out the
aggresive/passive optimization to save a factor of two in memory.  That
is, rather than computing win and notlose in the same pass, we compute
them in two separate passes.  The entries in our workspace change from
halfsupers_t to halfsuper_t, or from 32 to 16 bytes.

This results in some duplicate computation, but not too much since each
pass uses roughly half the memory bandwidth.  On my M2 MacBook Air:

  before: 2.25 s, 0.95 GB
  after:  3.10 s, 0.48 GB

And it works again on my iPhone 13 Pro: 7.2 s for 18 stones.